Refactor ReplicaManager (#813)
Remove BinlogReaderManger
Separate class SlaveNode into single file
